Moves setToken into it's own module
Also fixes path in fetchToken
